% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/songs.R
\docType{data}
\name{forever_for_now}
\alias{forever_for_now}
\title{The lyrics of LP's album "Forever For Now" (2014)}
\format{A list with 24 elements}
\source{
\url{https://genius.com/albums/Lp/Forever-for-now-deluxe-edition}
}
\usage{
forever_for_now
}
\description{
A dataset containing the lyrics of LP's 2014 album "Forever For Now".
The UTF-8 plain text was sourced manually created Excel files with the
lyrics obtained from Genius. (Some elements are blank.)
}
\keyword{datasets}
