\name{heart_shaped_scar}
\alias{heart_shaped_scar}
\docType{data}
\title{
Heart-Shaped Scar
}
\description{
Released 2001
}
\usage{data("heart_shaped_scar")}
\format{
  The format is:
List of 11
 $ perfect              : chr [1:35] "I got a nice ride you'd be perfect and any way we can work it" "It's a hard time for answers" "You could be my new dancer" "Eye candy and havoc" ...
 $ love_somebody        : chr [1:25] "She was a wild ,wild woman" "Always been that way" "She said, \"Baby, I want to go solo and you don't need me anyway\"" "" ...
 $ heart_shaped_scar    : chr [1:40] "I see a changing in the light" "A fire in the night" "You're better without me" "" ...
 $ coming_home          : chr [1:19] "When I look back I guess I only remember" "The good and sometimes the bad" "I hadn't seen you in a while Thought maybe you went off and had a child" "I was hoping I'd see you again" ...
 $ long_time            : chr [1:26] "I am homesick for my town" "Sometimes" "But I miss you all of the time" "By the way you can never stay" ...
 $ kiss_it_all_goodbye  : chr [1:31] "You think you're special" "You always have" "You're finding out hard core what its all about and now" "You're not so sure" ...
 $ shut_you_out         : chr [1:27] "Today's a brand new day" "You say let's get together" "You seem to think you're all I need" "Yeah, whatever" ...
 $ follow_me_down       : chr [1:42] "Burnt down cigarettes" "Faded blues and gray I'm feeling so worn" "I've seen better days" "You're so beautiful" ...
 $ insane               : chr [1:29] "Dark hallway" "Dishes in the sink" "The TVs on for day no sound" "Don't worry I'm losing sleep" ...
 $ when_i_m_not_with_you: chr [1:32] "I lie to you to get away" "Hide from myself to cure the pain" "I stamp my feet" "I cry out loud I keep you back" ...
 $ greenbriar           : chr [1:36] "I thought I saw young Jesus Christ" "Underneath the Greenbriar Ridge" "Holding hands one August night" "Not far from where we lived" ...
}
\source{
Genius (https://genius.com/albums/Lp/Heart-shaped-scar)
}
\examples{
heart_shaped_scar$perfect
}
\keyword{datasets}
