% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/songs.R
\docType{data}
\name{heart_to_mouth}
\alias{heart_to_mouth}
\title{The lyrics of LP's album "Heart to Mouth" (2018)}
\format{A list with 12 elements}
\source{
\url{https://genius.com/albums/Lp/Heart-to-mouth}
}
\usage{
heart_to_mouth
}
\description{
A dataset containing the lyrics of LP's 2018 album "Heart to Mouth".
The UTF-8 plain text was sourced manually created Excel files with the
lyrics obtained from Genius. (Some elements are blank.)
}
\keyword{datasets}
