% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/songs.R
\docType{data}
\name{lost_on_you}
\alias{lost_on_you}
\title{The lyrics of LP's album "Lost On You" (2016)}
\format{A list with 13 elements}
\source{
\url{https://genius.com/albums/Lp/Lost-on-you-international-version}
}
\usage{
lost_on_you
}
\description{
A dataset containing the lyrics of LP's 2016 album "Lost On You".
The UTF-8 plain text was sourced manually created Excel files with the
lyrics obtained from Genius. (Some elements are blank.)
}
\keyword{datasets}
