% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/songs.R
\docType{package}
\name{lp}
\alias{lp}
\alias{lp-package}
\title{LP's Lyrics}
\description{
This package contains the complete lyrics of LP's 5 completed,
published albums, formatted to be convenient for text analysis.
}
