% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/songs.R
\name{lp_songs}
\alias{lp_songs}
\title{Tidy data frame of LP's 5 completed, published albums}
\usage{
lp_songs()
}
\value{
A data frame with three columns: \code{album}, \code{song} and
\code{lyrics}
}
\description{
Returns a tidy data frame of LP's 5 completed, published albums with
three columns:  \code{album} which contains the titles of the albums as a
factor in order of publication, \code{song} which contains the titles of the
songs as factor in order of playing considering repetition, and
\code{lyrics}, which contains the lyrics of the songs.
}
\examples{
library(dplyr)

lp_songs() \%>\%
   group_by(album) \%>\%
   summarise(number_of_songs = n())

}
