% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/songs.R
\docType{data}
\name{suburban_sprawl_and_alcohol}
\alias{suburban_sprawl_and_alcohol}
\title{The lyrics of LP's album "Suburban Sprawl and Alcohol" (2004)}
\format{A list with 24 elements}
\source{
\url{https://genius.com/albums/Lp/Suburban-sprawl-alcohol}
}
\usage{
suburban_sprawl_and_alcohol
}
\description{
A dataset containing the lyrics of LP's 2001 album "Suburban Sprawl and Alcohol".
The UTF-8 plain text was sourced manually created Excel files with the
lyrics obtained from Genius. (Some elements are blank.)
}
\keyword{datasets}
