\name{forever_for_now}
\alias{forever_for_now}
\docType{data}
\title{
Forever For Now
}
\description{
Released January 26, 2018
}
\usage{data("forever_for_now")}
\format{
  The format is:
List of 24
 $ heavenly_light           : chr [1:81] "[Verse 1]" "In and outta doors" "In and outta wars" "Don't know what they're for" ...
 $ night_like_this          : chr [1:46] "[Chorus 1]" "Doo doo doo doo" "Doo doo doo" "Doo doo doo ooo" ...
 $ one_last_mistake         : chr [1:37] "[Verse 1]" "Ahead on the rise again, there's a road" "Suddenly something is up ahead, I see somebody I don't know" "Dreams are all I own, I guess I should go" ...
 $ tokyo_sunrise            : chr [1:45] "[Verse 1]" "Said a lot of words along the way" "I meant them all while we reigned" "But shores of love get beaten by the waves and after it was done I wish I'd saved time" ...
 $ salvation                : chr [1:46] "[Verse 1]" "When you arrived at my door I didn't recognize" "My salvation" "And the light, on the shore, it was in your eyes" ...
 $ your_town                : chr [1:60] "[Verse 1]" "Another year gone and lost and lately" "I would die just to talk or maybe hold you for a day" "It's not sunny like California" ...
 $ free_to_love             : chr [1:47] "[Verse 1]" "Oh say what you wanna say, nothing's ever gonna change" "Everybody's going away sometimes" "Bottom and monotony, virtual lobotomy" ...
 $ levitator                : chr [1:66] "[Verse 1]" "Baby I'd be lost today" "Like some wreckage from a hurricane" "Castaway" ...
 $ someday                  : chr [1:42] "[Verse 1]" "Some say when hearts race there's no saving grace" "Though cruel fate I'll just wait cus' nothing can sway" "" ...
 $ savannah                 : chr [1:68] "[Verse 1]" "The mountain was on fire" "And the southern night was dark" "As I drove into town" ...
 $ into_the_wild            : chr [1:59] "[Verse 1]" "Are we on the lonely side?" "Say oh-oh-oh now" "The past long away" ...
 $ forever_for_now          : chr [1:33] "[Verse 1]" "Hush, hush" "Don't say a word" "The faint cries can hardly be heard" ...
 $ los_in_the_world         : chr [1:50] "[Verse 1]" "The beautiful night and the cities of light" "You know they're priceless" "Millions of miles we've travelled" ...
 $ magic                    : chr [1:57] "[Verse 1]" "We paid our dues" "No lust for fame" "When tides get turned" ...
 $ road_to_ruin             : chr [1:74] "[Verse 1]" "I used to hear the silence" "I used to hear the chime" "Way down in the gold mine" ...
 $ into_the_wild_live       : chr [1:59] "[Verse 1]" "Are we on the lonely side?" "Say oh, oh, oh now" "The past long away" ...
 $ levitator_live           : chr [1:66] "[Verse 1]" "Baby I'd be lost today" "Like some wreckage from a hurricane" "Castaway" ...
 $ tokyo_sunrise_live       : chr [1:45] "[Verse 1]" "Said a lot of words along the way" "I meant them all while we reigned" "But shores of love get beaten by the waves and after it was done I wish I'd saved time" ...
 $ wasted_love_live         : chr [1:45] "I didn't want to love" "I was happy on my own" "I wish I never had a taste of your sweet song" "I was rolling like the stone" ...
 $ someday_live             : chr [1:42] "[Verse 1]" "Some say when hearts race there's no saving grace" "Though cruel fate I'll just wait cus' nothing can sway" "" ...
 $ fighting_with_myself_live: chr [1:101] "The battle has just begun" "You tell me I'm not the only one" "Cuz he's been loving you" "And you are all carried away" ...
 $ its_over_live            : chr [1:37] "[Intro]" "Your baby doesn't love you anymore" "" "[Verse 1]" ...
 $ halo_live                : chr [1:85] "Remember those walls I built" "Well, baby they're tumbling down" "And they didn't even put up a fight" "They didn't even make up a sound" ...
 $ forever_for_now_live     : chr [1:34] "[Introduction]" "*whistles*" "" "[Verse 1]" ...
}
\source{
Genius (https://genius.com/albums/Lp/Forever-for-now-deluxe-edition)
}
\examples{
forever_for_now$heavenly_light
}
\keyword{datasets}
