\name{heart_to_mouth}
\alias{heart_to_mouth}
\docType{data}
\title{
heart to Mouth
}
\description{
Released December 7, 2018
}
\usage{data("heart_to_mouth")}
\format{
  The format is:
List of 12
 $ dreamcatcher        : chr [1:31] "[Verse 1]" "Dreamcatcher, I can see you waving in the night alone" "Slow winter, and I feel I might get left out in the cold" "Ooh and I died today a little inside" ...
 $ when_i_m_over_you   : chr [1:70] "[Verse 1]" "Ooh, ooh" "Wanna fight, wanna get right" "Wanna live, wanna never die" ...
 $ one_night_in_the_sun: chr [1:62] "[Verse 1]" "Get your one night in the sun" "Be sure to take your adulation" "Get your pinch just when the moment's right" ...
 $ girls_go_wild       : chr [1:65] "[Verse 1]" "I've been caged, I've been hounded, I've been hunted and tamed" "I'm the outlaw of outside and ready to rage" "I'm in search of the heart I can eat to renew me, yeah" ...
 $ recovery            : chr [1:47] "[Verse 1]" "Our room is dark, the blinds are shut tight" "And everything is still too much outside" "It may be over, but not tonight" ...
 $ the_power           : chr [1:68] "[Intro]" "Woah" "Woah" "Yeah" ...
 $ dreamer             : chr [1:76] "[Intro]" "Baby, you're a dreamer" "You don't know what I need, yeah" "Now, I know you got heart" ...
 $ house_on_fire       : chr [1:47] "[Verse 1]" "I can't say yes or no" "We'll never know, just lay low" "The timing gets to me" ...
 $ hey_nice_to_know_ya : chr [1:67] "[Verse 1]" "It's been long, long drawn" "Got my heart broke in the car" "Parked outside the minibar" ...
 $ die_for_your_love   : chr [1:70] "[Verse 1]" "Turn the lights up, nothin' to hide from" "Let's see those eyes" "Turn your glance up, girl, just man up" ...
 $ shaken              : chr [1:70] "[Verse 1]" "All night, sippin' on a feelin'" "Won't lie, was nervous for this evenin'" "Then you arrive" ...
 $ special             : chr [1:60] "[Verse 1]" "Yeah, the money and the fame" "Is all bullshit in the end" "Brings you nothing else but pain" ...
}
\source{
Genius (https://genius.com/albums/Lp/Heart-to-mouth)
}
\examples{
heart_to_mouth$dreamcatcher
}
\keyword{datasets}
