\name{lost_on_you}
\alias{lost_on_you}
\docType{data}
\title{
Lost On You
}
\description{
Released May 5, 2017
}
\usage{data("lost_on_you")}
\format{
  The format is:
List of 10
 $ muddy_waters   : chr [1:53] "[Verse 1]" "We are kneeling at the rivers edge and tempting" "All the steps to follow closer right behind" "Is it only when you feel a part is empty" ...
 $ no_witness     : chr [1:65] "[Verse 1]" "How do I get you outta my head?" "How do I get you back in my bed?" "Oh no, oh no" ...
 $ lost_on_you    : chr [1:69] "[Verse 1]" "When you get older, plainer, saner" "Will you remember all the danger" "We came from?" ...
 $ up_against_me  : chr [1:59] "[Verse 1]" "I wanted more" "It went away" "I asked the Lord" ...
 $ tightrope      : chr [1:80] "[Verse 1]" "Hold on, I know you're scared" "But you're so close to heaven" "Eyes shut tight" ...
 $ other_people   : chr [1:52] "[Verse 1]" "I never tried to be a hero" "You took us from the stars to zero" "" ...
 $ into_the_wild  : chr [1:59] "[Verse 1]" "Are we on the lonely side?" "Say oh-oh-oh now" "The past long away" ...
 $ strange        : chr [1:52] "[Verse 1]" "Hey" "Anyone can get love" "Anyone can get laid" ...
 $ death_valley   : chr [1:54] "[Verse 1]" "Dime-store rock and roll" "Stretching out into forever" "I said, \"I don't know" ...
 $ you_want_it_all: chr [1:44] "[Verse 1]" "We been on this roundabout" "Swimming in these fields of doubt" "But there is no time now" ...
}
\source{
Genius (https://genius.com/albums/Lp/Lost-on-you-international-version)
}
\examples{
lost_on_you$muddy_waters
}
\keyword{datasets}
